feat(quality): Lot 12 quality cleanup - wave 4 (mypy, BaseWorker, tests, monolith split)
QUAL-011: Split veaf-tools.py (1517 lines) into veaf_tools/ package
  - veaf_tools/app.py: shared app, VERSION, constants, main_callback
  - veaf_tools/helpers.py: _read_single_char, _ask_replace, _update_build_config_in_yaml
  - veaf_tools/commands/ (11 modules, 14 commands): about, aircraft_groups, build,
    config, convert, convert_v5, extract, inject_presets, prepare, waypoints, weather
  - veaf-tools.py: reduced to thin ~25-line entry point

QUAL-012: BaseWorker ABC in veaf_libs/base_worker.py; applied to all 7 worker classes

QUAL-013: mypy fixes in mission_constants.py, miz_tools.py; removed 4 modules
  from ignore_errors list in pyproject.toml

QUAL-021: 16 pytest tests for miz_tools (all 16 passing); fixed test fixture
  to include theatre file by default

QUAL-022: dcs_mocks.lua enriched with configurable unit/group registry
  (addUnit, addGroup, removeUnit, removeGroup, clearUnitsAndGroups)

QUAL-023: TestVeafQraLifecycle (9 tests) + TestVeafQraUnit (4 tests) added
  to test_veafQraManager.lua
