+++ pyproject.toml
@@ -103,12 +103,9 @@ module = [
     "mission_builder.mission_builder_worker",
     "mission_converter.mission_converter_worker",
     "mission_extractor.mission_extractor_worker",
-    "mission_tools.mission_constants",
-    "mission_tools.miz_tools",
     "presets_injector.presets_injector_worker",
     "presets_injector.presets_manager",
     "presets_injector.test_presets",
-    "veaf_libs.logger",
     "waypoints_injector.waypoints_injector_worker",
     "waypoints_injector.waypoints_manager",
     "weather_injector.utils.lua_converter",
