+++ src/python/veaf-tools/veaf_tools/commands/__init__.py
@@ -1,0 +1,1 @@
+from . import about, aircraft_groups, build, config, convert, convert_v5, extract, inject_presets, prepare, waypoints, weather  # noqa: F401
