+++ src/python/veaf-tools/veaf_tools/commands/extract.py
@@ -1,0 +1,66 @@
+from pathlib import Path
+
+import typer
+from mission_extractor import MissionExtractorREADME, MissionExtractorWorker
+from rich.markdown import Markdown
+from veaf_libs.paths import resolve_path
+
+from veaf_tools.app import (
+    DEFAULT_MISSION_FILE,
+    PAUSE_HELP,
+    README_HELP,
+    VERBOSE_HELP,
+    VERSION,
+    app,
+    console,
+    logger,
+    t,
+)
+
+
+@app.command(no_args_is_help=True)
+def extract(
+    readme: bool = typer.Option(False, help=README_HELP),
+    verbose: bool = typer.Option(False, help=VERBOSE_HELP),
+    mission_name_or_file: str | None = typer.Argument(
+        DEFAULT_MISSION_FILE,
+        help="Mission name; will extract from the mission with this name (most recent .miz file); can be set to a .miz file.",
+    ),
+    mission_folder: str | None = typer.Argument(".", help="Folder where the mission files will be extracted."),
+    pause: bool = typer.Option(False, help=PAUSE_HELP),
+) -> None:
+    """
+    Extracts a DCS mission .miz file to a VEAF mission folder.
+    """
+
+    logger.set_verbose(verbose)
+
+    # Set the title and version
+    console.print(f"[bold green]veaf-tools VEAF mission extractor v{VERSION}[/bold green]")
+
+    if readme:
+        if typer.confirm(t("help.confirm_doc")):
+            md_render = Markdown(MissionExtractorREADME)
+            console.print(md_render)
+        exit()
+
+    # Resolve output mission folder
+    p_mission_folder = resolve_path(path=mission_folder, default_path=Path.cwd(), create_if_not_exist=True)
+    if not p_mission_folder.exists():
+        logger.error(f"Mission folder {p_mission_folder} does not exist!", exception_type=FileNotFoundError)
+
+    # Resolve input mission
+    assert mission_name_or_file is not None
+    p_input_mission: str | Path | None = mission_name_or_file
+    if not mission_name_or_file.lower().endswith(".miz"):
+        if files := list(p_mission_folder.glob(f"{mission_name_or_file}*.miz")):
+            p_input_mission = max(files, key=lambda f: f.stat().st_mtime)
+    p_input_mission = resolve_path(path=p_input_mission, should_exist=True)
+
+    # Call the worker class
+    worker = MissionExtractorWorker(mission_folder=p_mission_folder, input_mission_path=p_input_mission)
+    worker.work()
+
+    console.print(t("msg.work_done"))
+    if pause:
+        input(t("help.pause_msg"))
